refactor: passive is job specific resource
