+++ simaple/data/__init__.py
@@ -1,6 +1,5 @@
 from simaple.data.ability import get_best_ability
 
-
 __all__ = [
     "get_best_ability",
 ]
