+++ simaple/data/jobs/__init__.py
@@ -1,4 +1,4 @@
-from simaple.data.jobs.builtin import get_skill_profile, get_every_hyper_skills
+from simaple.data.jobs.builtin import get_every_hyper_skills, get_skill_profile
 
 __all__ = [
     "get_skill_profile",
