+++ simaple/data/jobs/definitions/__init__.py
@@ -1,5 +1,10 @@
+from simaple.data.jobs.definitions.passive_hyper_skill import (
+    MultiplierPassiveHyperskill,
+    PassiveHyperskillInterface,
+    StatIncreasePassiveHyperskill,
+    ValueIncreasePassiveHyperskill,
+)
 from simaple.data.jobs.definitions.skill_profile import SkillProfile
-from simaple.data.jobs.definitions.passive_hyper_skill import PassiveHyperskillInterface, MultiplierPassiveHyperskill, StatIncreasePassiveHyperskill, ValueIncreasePassiveHyperskill
 
 __all__ = [
     "SkillProfile",
