+++ simaple/data/jobs/resources/passive_skill/adele.yaml
