+++ simaple/data/jobs/resources/passive_skill/archmagefb.yaml
