+++ simaple/data/jobs/resources/passive_skill/archmagetc.yaml
