+++ simaple/data/jobs/resources/passive_skill/bishop.yaml
