+++ simaple/data/jobs/resources/passive_skill/dualblade.yaml
