+++ simaple/data/jobs/resources/passive_skill/mechanic.yaml
