+++ simaple/data/jobs/resources/passive_skill/soulmaster.yaml
