+++ simaple/data/jobs/resources/passive_skill/thief.yaml
