+++ simaple/data/jobs/resources/passive_skill/windbreaker.yaml
